separate out gear picker components
