+++ ui/core/components/gear_picker/gear_picker.tsx
@@ -1,58 +1,51 @@
-import clsx from 'clsx';
-import tippy from 'tippy.js';
 import { ref } from 'tsx-vanilla';
 
-import { SortDirection } from '../../constants/other';
 import { setItemQualityCssClass } from '../../css_utils';
-import { IndividualSimUI } from '../../individual_sim_ui';
-import { Player, ReforgeData } from '../../player';
-import { Class, GemColor, ItemQuality, ItemRandomSuffix, ItemSlot, ItemSpec, ItemType } from '../../proto/common';
-import { DatabaseFilters, RepFaction, UIEnchant as Enchant, UIGem as Gem, UIItem as Item, UIItem_FactionRestriction } from '../../proto/ui';
+import { Player } from '../../player';
+import { GemColor, ItemSlot, ItemType } from '../../proto/common';
+import { UIEnchant as Enchant, UIGem as Gem } from '../../proto/ui';
 import { ActionId } from '../../proto_utils/action_id';
-import { getEnchantDescription, getUniqueEnchantString } from '../../proto_utils/enchants';
+import { getEnchantDescription } from '../../proto_utils/enchants';
 import { EquippedItem } from '../../proto_utils/equipped_item';
-import { gemMatchesSocket, getEmptyGemSocketIconUrl } from '../../proto_utils/gems';
-import {
-	difficultyNames,
-	professionNames,
-	REP_FACTION_NAMES,
-	REP_FACTION_QUARTERMASTERS,
-	REP_LEVEL_NAMES,
-	shortSecondaryStatNames,
-	slotNames,
-} from '../../proto_utils/names';
-import { Stats } from '../../proto_utils/stats';
-import { getPVPSeasonFromItem, isPVPItem } from '../../proto_utils/utils';
-import { Sim } from '../../sim';
+import { getEmptyGemSocketIconUrl } from '../../proto_utils/gems';
+import { shortSecondaryStatNames, slotNames } from '../../proto_utils/names';
 import { SimUI } from '../../sim_ui';
-import { EventID, TypedEvent } from '../../typed_event';
-import { formatDeltaTextElem, mod, randomUUID, sanitizeId } from '../../utils';
-import { BaseModal } from '../base_modal';
+import { EventID } from '../../typed_event';
 import { Component } from '../component';
-import { FiltersMenu } from '../filters_menu';
-import {
-	makePhaseSelector,
-	makeShow1hWeaponsSelector,
-	makeShow2hWeaponsSelector,
-	makeShowEPValuesSelector,
-	makeShowMatchingGemsSelector,
-} from '../other_inputs';
 import QuickSwapList from '../quick_swap';
-import Toast from '../toast';
-import { Clusterize } from '../virtual_scroll/clusterize';
+import { GearData } from './item_list';
 import { addQuickEnchantPopover } from './quick_enchant_popover';
 import { addQuickGemPopover } from './quick_gem_popover';
+import SelectorModal, { SelectorModalTabs } from './selector_modal';
 
-const EP_TOOLTIP = `
-	EP (Equivalence Points) is way of comparing items by multiplying the raw stats of an item with your current stat weights.
-	More EP does not necessarily mean more DPS, as EP doesn't take into account stat caps and non-linear stat calculations.
-`;
+const emptySlotIcons: Record<ItemSlot, string> = {
+	[ItemSlot.ItemSlotHead]: '/cata/assets/item_slots/head.jpg',
+	[ItemSlot.ItemSlotNeck]: '/cata/assets/item_slots/neck.jpg',
+	[ItemSlot.ItemSlotShoulder]: '/cata/assets/item_slots/shoulders.jpg',
+	[ItemSlot.ItemSlotBack]: '/cata/assets/item_slots/shirt.jpg',
+	[ItemSlot.ItemSlotChest]: '/cata/assets/item_slots/chest.jpg',
+	[ItemSlot.ItemSlotWrist]: '/cata/assets/item_slots/wrists.jpg',
+	[ItemSlot.ItemSlotHands]: '/cata/assets/item_slots/hands.jpg',
+	[ItemSlot.ItemSlotWaist]: '/cata/assets/item_slots/waist.jpg',
+	[ItemSlot.ItemSlotLegs]: '/cata/assets/item_slots/legs.jpg',
+	[ItemSlot.ItemSlotFeet]: '/cata/assets/item_slots/feet.jpg',
+	[ItemSlot.ItemSlotFinger1]: '/cata/assets/item_slots/finger.jpg',
+	[ItemSlot.ItemSlotFinger2]: '/cata/assets/item_slots/finger.jpg',
+	[ItemSlot.ItemSlotTrinket1]: '/cata/assets/item_slots/trinket.jpg',
+	[ItemSlot.ItemSlotTrinket2]: '/cata/assets/item_slots/trinket.jpg',
+	[ItemSlot.ItemSlotMainHand]: '/cata/assets/item_slots/mainhand.jpg',
+	[ItemSlot.ItemSlotOffHand]: '/cata/assets/item_slots/offhand.jpg',
+	[ItemSlot.ItemSlotRanged]: '/cata/assets/item_slots/ranged.jpg',
+};
+export function getEmptySlotIconUrl(slot: ItemSlot): string {
+	return emptySlotIcons[slot];
+}
 
-const createHeroicLabel = () => {
+export const createHeroicLabel = () => {
 	return <span className="heroic-label">[H]</span>;
 };
 
-const createGemContainer = (socketColor: GemColor, gem: Gem | null, index: number) => {
+export const createGemContainer = (socketColor: GemColor, gem: Gem | null, index: number) => {
 	const gemIconElem = ref<HTMLImageElement>();
 	const gemContainerElem = ref<HTMLAnchorElement>();
 	const gemContainer = (
@@ -457,1354 +450,5 @@ export class IconItemSwapPicker extends Component {
 			getEquippedItem: () => this.player.getItemSwapItem(this.slot),
 			changeEvent: this.player.itemSwapChangeEmitter,
 		};
-	}
-}
-
-export interface GearData {
-	equipItem: (eventID: EventID, equippedItem: EquippedItem | null) => void;
-	getEquippedItem: () => EquippedItem | null;
-	changeEvent: TypedEvent<any>;
-}
-
-export enum SelectorModalTabs {
-	Items = 'Items',
-	RandomSuffixes = 'Random Suffix',
-	Enchants = 'Enchants',
-	Reforging = 'Reforging',
-	Gem1 = 'Gem1',
-	Gem2 = 'Gem2',
-	Gem3 = 'Gem3',
-}
-
-type SelectorModalOptions = {
-	// This will add a unique ID to the modal, allowing multiple of the same modals to exist
-	id: string;
-	// Prevents rendering of certail tabs
-	disabledTabs?: SelectorModalTabs[];
-	// If defined will render a remove button in the header
-	removeButtonText?: string;
-};
-export class SelectorModal extends BaseModal {
-	private readonly simUI: SimUI;
-	private player: Player<any>;
-	private gearPicker: GearPicker | undefined;
-	private ilists: ItemList<ItemListType>[] = [];
-
-	private readonly itemSlotTabElems: HTMLElement[] = [];
-	private readonly titleElem: HTMLElement;
-	private readonly tabsElem: HTMLElement;
-	private readonly contentElem: HTMLElement;
-	public readonly removeButton?: HTMLButtonElement;
-
-	private currentSlot: ItemSlot = ItemSlot.ItemSlotHead;
-	private currentTab: SelectorModalTabs = SelectorModalTabs.Items;
-	private disabledTabs: SelectorModalTabs[] = [];
-	private options: SelectorModalOptions;
-
-	constructor(parent: HTMLElement, simUI: SimUI, player: Player<any>, gearPicker?: GearPicker, options?: Partial<SelectorModalOptions>) {
-		super(parent, 'selector-modal', { disposeOnClose: false, size: 'xl' });
-
-		this.simUI = simUI;
-		this.player = player;
-		this.gearPicker = gearPicker;
-		this.options = { id: randomUUID(), ...options };
-		this.disabledTabs = this.options.disabledTabs || [];
-
-		this.addItemSlotTabs();
-
-		this.header!.insertAdjacentElement(
-			'afterbegin',
-			<div>
-				<h6 className="selector-modal-title" />
-				<ul className="nav nav-tabs selector-modal-tabs"></ul>
-			</div>,
-		);
-
-		if (this.options.removeButtonText) {
-			const closeX = this.header?.querySelector('.btn-close');
-			if (!!closeX) {
-				const removeButtonRef = ref<HTMLButtonElement>();
-				this.header?.insertBefore(
-					<button ref={removeButtonRef} className="btn btn-danger">
-						{this.options.removeButtonText}
-					</button>,
-					closeX,
-				);
-				if (removeButtonRef.value) this.removeButton = removeButtonRef.value;
-			}
-		}
-
-		this.body.appendChild(<div className="tab-content selector-modal-tab-content"></div>);
-
-		this.titleElem = this.rootElem.querySelector<HTMLElement>('.selector-modal-title')!;
-		this.tabsElem = this.rootElem.querySelector<HTMLElement>('.selector-modal-tabs')!;
-		this.contentElem = this.rootElem.querySelector<HTMLElement>('.selector-modal-tab-content')!;
-
-		this.body.appendChild(
-			<div className="d-flex align-items-center form-text mt-3">
-				<i className="fas fa-circle-exclamation fa-xl me-2"></i>
-				<span>
-					If gear is missing, check the selected phase and your gear filters.
-					<br />
-					If the problem persists, save any un-saved data, click the
-					<i className="fas fa-cog mx-1"></i>
-					to open your sim options, then click the "Restore Defaults".
-				</span>
-			</div>,
-		);
-	}
-
-	openTab(selectedSlot: ItemSlot, selectedTab: SelectorModalTabs, gearData: GearData) {
-		this.titleElem.textContent = slotNames.get(selectedSlot) ?? '';
-		this.setData(selectedSlot, selectedTab, gearData);
-		this.setActiveItemSlotTab(selectedSlot);
-		this.open();
-	}
-
-	onShow() {
-		if (this.gearPicker) {
-			// Allow you to switch between gear picker slots with the up and down arrows
-			const switchToPreviousItemSlotTab = this.switchToPreviousItemSlotTab.bind(this);
-			const switchToNextItemSlotTab = this.switchToNextItemSlotTab.bind(this);
-
-			document.addEventListener('keydown', switchToPreviousItemSlotTab);
-			document.addEventListener('keydown', switchToNextItemSlotTab);
-
-			this.addOnHideCallback(() => document.removeEventListener('keydown', switchToPreviousItemSlotTab));
-			this.addOnHideCallback(() => document.removeEventListener('keydown', switchToNextItemSlotTab));
-		}
-	}
-
-	private setData(selectedSlot: ItemSlot, selectedTab: SelectorModalTabs, gearData: GearData) {
-		this.tabsElem.innerText = '';
-		this.contentElem.innerText = '';
-		this.ilists = [];
-
-		const equippedItem = gearData.getEquippedItem();
-
-		const eligibleItems = this.player.getItems(selectedSlot);
-		const eligibleEnchants = this.player.getEnchants(selectedSlot);
-		const eligibleReforges = equippedItem?.item ? this.player.getAvailableReforgings(equippedItem.getWithRandomSuffixStats()) : [];
-		// If the enchant tab is selected but the item has no eligible enchants, default to items
-		// If the reforge tab is selected but the item has no eligible reforges, default to items
-		// If a gem tab is selected but the item has no eligible sockets, default to items
-		if (
-			(selectedTab === SelectorModalTabs.Enchants && !eligibleEnchants.length) ||
-			(selectedTab === SelectorModalTabs.Reforging && !eligibleReforges.length) ||
-			([SelectorModalTabs.Gem1, SelectorModalTabs.Gem2, SelectorModalTabs.Gem3].includes(selectedTab) &&
-				equippedItem?.numSockets(this.player.isBlacksmithing()) === 0)
-		) {
-			selectedTab = SelectorModalTabs.Items;
-		}
-
-		this.currentTab = selectedTab;
-		this.currentSlot = selectedSlot;
-
-		const hasItemTab = !this.disabledTabs?.includes(SelectorModalTabs.Items);
-		if (hasItemTab)
-			this.addTab<Item>({
-				id: sanitizeId(`${this.options.id}-${SelectorModalTabs.Items}`),
-				label: SelectorModalTabs.Items,
-				gearData,
-				itemData: eligibleItems.map(item => {
-					return {
-						item: item,
-						id: item.id,
-						actionId: ActionId.fromItem(item),
-						name: item.name,
-						quality: item.quality,
-						heroic: item.heroic,
-						phase: item.phase,
-						baseEP: this.player.computeItemEP(item, selectedSlot),
-						ignoreEPFilter: false,
-						onEquip: (eventID, item) => {
-							const equippedItem = gearData.getEquippedItem();
-							if (equippedItem) {
-								gearData.equipItem(eventID, equippedItem.withItem(item));
-							} else {
-								gearData.equipItem(eventID, new EquippedItem(item));
-							}
-						},
-					};
-				}),
-				computeEP: (item: Item) => this.player.computeItemEP(item, selectedSlot),
-				equippedToItemFn: (equippedItem: EquippedItem | null) => equippedItem?.item,
-				onRemove: (eventID: number) => {
-					gearData.equipItem(eventID, null);
-					this.removeTabs('Gem');
-					this.removeTabs(SelectorModalTabs.RandomSuffixes);
-				},
-			});
-
-		const hasEnchantTab = !this.disabledTabs?.includes(SelectorModalTabs.Enchants);
-		if (hasEnchantTab)
-			this.addTab<Enchant>({
-				id: sanitizeId(`${this.options.id}-${SelectorModalTabs.Enchants}`),
-				label: SelectorModalTabs.Enchants,
-				gearData,
-				itemData: eligibleEnchants.map(enchant => {
-					return {
-						item: enchant,
-						id: enchant.effectId,
-						actionId: enchant.itemId ? ActionId.fromItemId(enchant.itemId) : ActionId.fromSpellId(enchant.spellId),
-						name: enchant.name,
-						quality: enchant.quality,
-						phase: enchant.phase || 1,
-						baseEP: this.player.computeStatsEP(new Stats(enchant.stats)),
-						ignoreEPFilter: true,
-						heroic: false,
-						onEquip: (eventID, enchant) => {
-							const equippedItem = gearData.getEquippedItem();
-							if (equippedItem) gearData.equipItem(eventID, equippedItem.withEnchant(enchant));
-						},
-					};
-				}),
-				computeEP: (enchant: Enchant) => this.player.computeEnchantEP(enchant),
-				equippedToItemFn: (equippedItem: EquippedItem | null) => equippedItem?.enchant,
-				onRemove: (eventID: number) => {
-					const equippedItem = gearData.getEquippedItem();
-					if (equippedItem) gearData.equipItem(eventID, equippedItem.withEnchant(null));
-				},
-			});
-
-		const hasRandomSuffixTab = !this.disabledTabs?.includes(SelectorModalTabs.RandomSuffixes);
-		if (hasRandomSuffixTab) this.addRandomSuffixTab(equippedItem, gearData);
-		const hasReforgingTab = !this.disabledTabs?.includes(SelectorModalTabs.Reforging);
-		if (hasReforgingTab) this.addReforgingTab(equippedItem, gearData);
-		const hasGemsTab = ![SelectorModalTabs.Gem1, SelectorModalTabs.Gem2, SelectorModalTabs.Gem3].some(gem => this.disabledTabs?.includes(gem));
-		if (hasGemsTab) this.addGemTabs(selectedSlot, equippedItem, gearData);
-
-		this.ilists.find(list => selectedTab === list.label)?.sizeRefresh();
-	}
-
-	private addItemSlotTabs() {
-		if (!this.gearPicker) {
-			return;
-		}
-
-		this.dialog.prepend(
-			<div className="gear-picker-modal-slots">
-				{this.gearPicker.itemPickers.map(picker => {
-					const anchorRef = ref<HTMLAnchorElement>();
-					const wrapper = (
-						<div className="item-picker-icon-wrapper" dataset={{ slot: picker.slot }}>
-							<a
-								ref={anchorRef}
-								className="item-picker-icon"
-								href="javascript:void(0)"
-								onclick={(e: Event) => {
-									e.preventDefault();
-									if (picker.slot != this.currentSlot) {
-										picker.openSelectorModal(this.currentTab);
-									}
-								}}
-								dataset={{ whtticon: 'false' }}
-							/>
-						</div>
-					) as HTMLElement;
-
-					const setItemData = () => {
-						if (picker.item) {
-							this.player.setWowheadData(picker.item, anchorRef.value!);
-							picker.item
-								.asActionId()
-								.fill()
-								.then(filledId => {
-									filledId.setBackgroundAndHref(anchorRef.value!);
-								});
-						} else {
-							anchorRef.value!.style.backgroundImage = `url('${getEmptySlotIconUrl(picker.slot)}')`;
-						}
-					};
-					setItemData();
-					picker.onUpdate(() => setItemData());
-					tippy(anchorRef.value!, {
-						content: `Edit ${slotNames.get(picker.slot)}`,
-						placement: 'left',
-					});
-					this.itemSlotTabElems.push(wrapper);
-					return wrapper;
-				})}
-			</div>,
-		);
-	}
-
-	private setActiveItemSlotTab(slot: ItemSlot) {
-		this.itemSlotTabElems.forEach(elem => {
-			if (elem.dataset.slot === slot.toString()) {
-				elem.classList.add('active');
-			} else if (elem.classList.contains('active')) {
-				elem.classList.remove('active');
-			}
-		});
-	}
-
-	private switchToPreviousItemSlotTab(event: KeyboardEvent) {
-		if (event.key === 'ArrowUp' && this.gearPicker) {
-			event.preventDefault();
-			const newSlot = mod(this.currentSlot - 1, Object.keys(ItemSlot).length / 2) as unknown as ItemSlot;
-			this.gearPicker.itemPickers[newSlot].openSelectorModal(this.currentTab);
-		}
-	}
-
-	private switchToNextItemSlotTab(event: KeyboardEvent) {
-		if (event.key === 'ArrowDown' && this.gearPicker) {
-			event.preventDefault();
-			const newSlot = mod(this.currentSlot + 1, Object.keys(ItemSlot).length / 2) as unknown as ItemSlot;
-			this.gearPicker.itemPickers[newSlot].openSelectorModal(this.currentTab);
-		}
-	}
-
-	private addGemTabs(_slot: ItemSlot, equippedItem: EquippedItem | null, gearData: GearData) {
-		if (!equippedItem) {
-			return;
-		}
-
-		const socketBonusEP = this.player.computeStatsEP(new Stats(equippedItem.item.socketBonus)) / (equippedItem.item.gemSockets.length || 1);
-		equippedItem.curSocketColors(this.player.isBlacksmithing()).forEach((socketColor, socketIdx) => {
-			const label = SelectorModalTabs[`Gem${socketIdx + 1}` as keyof typeof SelectorModalTabs];
-			this.addTab<Gem>({
-				id: sanitizeId(`${this.options.id}-${label}`),
-				label,
-				gearData,
-				itemData: this.player.getGems(socketColor).map((gem: Gem) => {
-					return {
-						item: gem,
-						id: gem.id,
-						actionId: ActionId.fromItemId(gem.id),
-						name: gem.name,
-						quality: gem.quality,
-						phase: gem.phase,
-						heroic: false,
-						baseEP: this.player.computeStatsEP(new Stats(gem.stats)),
-						ignoreEPFilter: true,
-						onEquip: (eventID, gem) => {
-							const equippedItem = gearData.getEquippedItem();
-							if (equippedItem) gearData.equipItem(eventID, equippedItem.withGem(gem, socketIdx));
-						},
-					};
-				}),
-				computeEP: (gem: Gem) => {
-					let gemEP = this.player.computeGemEP(gem);
-					if (gemMatchesSocket(gem, socketColor)) {
-						gemEP += socketBonusEP;
-					}
-					return gemEP;
-				},
-				equippedToItemFn: (equippedItem: EquippedItem | null) => equippedItem?.gems[socketIdx],
-				onRemove: (eventID: number) => {
-					const equippedItem = gearData.getEquippedItem();
-					if (equippedItem) gearData.equipItem(eventID, equippedItem.withGem(null, socketIdx));
-				},
-				setTabContent: (tabAnchor: HTMLAnchorElement) => {
-					const gemContainer = createGemContainer(socketColor, null, socketIdx);
-					tabAnchor.appendChild(gemContainer);
-					tabAnchor.classList.add('selector-modal-tab-gem');
-
-					const gemElem = tabAnchor.querySelector<HTMLElement>('.gem-icon')!;
-					const emptySocketUrl = getEmptyGemSocketIconUrl(socketColor);
-
-					const updateGemIcon = () => {
-						const equippedItem = gearData.getEquippedItem();
-						const gem = equippedItem?.gems[socketIdx];
-
-						if (gem) {
-							gemElem.classList.remove('hide');
-							ActionId.fromItemId(gem.id)
-								.fill()
-								.then(filledId => {
-									gemElem.setAttribute('src', filledId.iconUrl);
-								});
-						} else {
-							gemElem.classList.add('hide');
-							gemElem.setAttribute('src', emptySocketUrl);
-						}
-					};
-
-					gearData.changeEvent.on(updateGemIcon);
-					this.addOnDisposeCallback(() => gearData.changeEvent.off(updateGemIcon));
-					updateGemIcon();
-				},
-				socketColor,
-			});
-		});
-	}
-
-	private addRandomSuffixTab(equippedItem: EquippedItem | null, gearData: GearData) {
-		if (!equippedItem || !equippedItem.item.randomSuffixOptions.length) {
-			return;
-		}
-
-		const itemProto = equippedItem.item;
-
-		this.addTab<ItemRandomSuffix>({
-			id: sanitizeId(`${this.options.id}-${SelectorModalTabs.RandomSuffixes}`),
-			label: SelectorModalTabs.RandomSuffixes,
-			gearData,
-			itemData: this.player.getRandomSuffixes(itemProto).map((randomSuffix: ItemRandomSuffix) => {
-				return {
-					item: randomSuffix,
-					id: randomSuffix.id,
-					actionId: ActionId.fromRandomSuffix(itemProto, randomSuffix),
-					name: randomSuffix.name,
-					quality: itemProto.quality,
-					phase: itemProto.phase,
-					heroic: false,
-					baseEP: this.player.computeRandomSuffixEP(randomSuffix),
-					ignoreEPFilter: true,
-					onEquip: (eventID, randomSuffix) => {
-						const equippedItem = gearData.getEquippedItem();
-						if (equippedItem) gearData.equipItem(eventID, equippedItem.withItem(equippedItem.item).withRandomSuffix(randomSuffix));
-					},
-				};
-			}),
-			computeEP: (randomSuffix: ItemRandomSuffix) => this.player.computeRandomSuffixEP(randomSuffix),
-			equippedToItemFn: (equippedItem: EquippedItem | null) => equippedItem?.randomSuffix,
-			onRemove: (eventID: number) => {
-				const equippedItem = gearData.getEquippedItem();
-				if (equippedItem) gearData.equipItem(eventID, equippedItem.withRandomSuffix(null));
-			},
-		});
-	}
-
-	private addReforgingTab(equippedItem: EquippedItem | null, gearData: GearData) {
-		if (!equippedItem || (equippedItem.hasRandomSuffixOptions() && !equippedItem.randomSuffix)) {
-			return;
-		}
-
-		const itemProto = equippedItem.item;
-
-		this.addTab<ReforgeData>({
-			id: sanitizeId(`${this.options.id}-${SelectorModalTabs.Reforging}`),
-			label: SelectorModalTabs.Reforging,
-			gearData,
-			itemData: this.player.getAvailableReforgings(equippedItem).map(reforgeData => {
-				return {
-					item: reforgeData,
-					id: reforgeData.id,
-					actionId: ActionId.fromReforge(itemProto, reforgeData.reforge),
-					name: (
-						<div>
-							<span className="reforge-value negative">
-								{reforgeData.fromAmount} {shortSecondaryStatNames.get(reforgeData.fromStat[0])}
-							</span>
-							<span className="reforge-value positive">
-								+{reforgeData.toAmount} {shortSecondaryStatNames.get(reforgeData.toStat[0])}
-							</span>
-						</div>
-					) as HTMLElement,
-					quality: ItemQuality.ItemQualityCommon,
-					phase: itemProto.phase,
-					heroic: false,
-					baseEP: this.player.computeReforgingEP(reforgeData),
-					ignoreEPFilter: true,
-					onEquip: (eventID, reforgeData) => {
-						const equippedItem = gearData.getEquippedItem();
-						if (equippedItem) gearData.equipItem(eventID, equippedItem.withReforge(reforgeData.reforge));
-					},
-				};
-			}),
-			computeEP: (reforge: ReforgeData) => this.player.computeReforgingEP(reforge),
-			equippedToItemFn: (equippedItem: EquippedItem | null) =>
-				equippedItem?.reforge ? this.player.getReforgeData(equippedItem, equippedItem.reforge) : null,
-			onRemove: (eventID: number) => {
-				const equippedItem = gearData.getEquippedItem();
-				if (equippedItem) gearData.equipItem(eventID, equippedItem.withItem(equippedItem.item));
-			},
-		});
-	}
-
-	/**
-	 * Adds one of the tabs for the item selector menu.
-	 *
-	 * T is expected to be Item, Enchant, or Gem. Tab menus for all 3 looks extremely
-	 * similar so this function uses extra functions to do it generically.
-	 */
-	private addTab<T extends ItemListType>({
-		id,
-		label,
-		gearData,
-		itemData,
-		computeEP,
-		equippedToItemFn,
-		onRemove,
-		setTabContent,
-		socketColor,
-	}: {
-		id: string;
-		label: SelectorModalTabs;
-		gearData: GearData;
-		itemData: ItemData<T>[];
-		computeEP: (item: T) => number;
-		equippedToItemFn: (equippedItem: EquippedItem | null) => T | null | undefined;
-		onRemove: (eventID: EventID) => void;
-		setTabContent?: (tabElem: HTMLAnchorElement) => void;
-		socketColor?: GemColor;
-	}) {
-		if (!itemData.length) {
-			return;
-		}
-		const selected = label === this.currentTab;
-		const tabAnchor = ref<HTMLAnchorElement>();
-		this.tabsElem.appendChild(
-			<li className="nav-item">
-				<a
-					ref={tabAnchor}
-					className={clsx('nav-link selector-modal-item-tab', selected && 'active')}
-					dataset={{
-						label,
-						bsToggle: 'tab',
-						bsTarget: `#${id}`,
-					}}
-					attributes={{
-						role: 'tab',
-						'aria-selected': selected,
-					}}></a>
-			</li>,
-		);
-
-		if (setTabContent) {
-			setTabContent(tabAnchor.value!);
-		} else {
-			tabAnchor.value!.textContent = label;
-		}
-
-		const ilist = new ItemList(
-			id,
-			this.contentElem,
-			this.simUI,
-			this.currentSlot,
-			this.currentTab,
-			this.player,
-			label,
-			gearData,
-			itemData,
-			socketColor || GemColor.GemColorUnknown,
-			computeEP,
-			equippedToItemFn,
-			onRemove,
-			itemData => {
-				const item = itemData;
-				itemData.onEquip(TypedEvent.nextEventID(), item.item);
-
-				const isItemChange = Item.is(item.item);
-				const isRandomSuffixChange = !!item.actionId.randomSuffixId;
-				// If the item changes, then gem slots and random suffix options will also change, so remove and recreate these tabs.
-				if (isItemChange || isRandomSuffixChange) {
-					if (!isRandomSuffixChange) {
-						this.removeTabs(SelectorModalTabs.RandomSuffixes);
-						this.addRandomSuffixTab(gearData.getEquippedItem(), gearData);
-					}
-
-					this.removeTabs(SelectorModalTabs.Reforging);
-					this.addReforgingTab(gearData.getEquippedItem(), gearData);
-
-					this.removeTabs('Gem');
-					this.addGemTabs(this.currentSlot, gearData.getEquippedItem(), gearData);
-				}
-			},
-		);
-
-		const invokeUpdate = () => {
-			ilist.updateSelected();
-		};
-		const applyFilter = () => {
-			ilist.applyFilters();
-		};
-		const hideOrShowEPValues = () => {
-			ilist.hideOrShowEPValues();
-		};
-		// Add event handlers
-		gearData.changeEvent.on(invokeUpdate);
-
-		this.player.sim.phaseChangeEmitter.on(applyFilter);
-		this.player.sim.filtersChangeEmitter.on(applyFilter);
-		this.player.sim.showEPValuesChangeEmitter.on(hideOrShowEPValues);
-
-		this.addOnDisposeCallback(() => {
-			gearData.changeEvent.off(invokeUpdate);
-			this.player.sim.phaseChangeEmitter.off(applyFilter);
-			this.player.sim.filtersChangeEmitter.off(applyFilter);
-			this.player.sim.showEPValuesChangeEmitter.off(hideOrShowEPValues);
-			ilist.dispose();
-		});
-
-		tabAnchor.value!.addEventListener('click', _event => {
-			this.currentTab = label;
-		});
-		tabAnchor.value!.addEventListener('shown.bs.tab', _event => {
-			ilist.sizeRefresh();
-		});
-
-		this.ilists.push(ilist as unknown as ItemList<ItemListType>);
-	}
-
-	private removeTabs(labelSubstring: string) {
-		const tabElems = [...this.tabsElem.querySelectorAll<HTMLElement>('.selector-modal-item-tab')].filter(
-			tab => tab.dataset?.label?.includes(labelSubstring),
-		);
-
-		const contentElems = tabElems.map(tabElem => document.querySelector(tabElem.dataset.bsTarget!)).filter(tabElem => Boolean(tabElem));
-		tabElems.forEach(elem => elem.parentElement?.remove());
-		contentElems.forEach(elem => elem!.remove());
-	}
-}
-
-export interface ItemData<T extends ItemListType> {
-	item: T;
-	name: string | HTMLElement;
-	id: number;
-	actionId: ActionId;
-	quality: ItemQuality;
-	phase: number;
-	baseEP: number;
-	ignoreEPFilter: boolean;
-	heroic: boolean;
-	onEquip: (eventID: EventID, item: T) => void;
-}
-
-interface ItemDataWithIdx<T extends ItemListType> {
-	idx: number;
-	data: ItemData<T>;
-}
-
-const emptySlotIcons: Record<ItemSlot, string> = {
-	[ItemSlot.ItemSlotHead]: '/cata/assets/item_slots/head.jpg',
-	[ItemSlot.ItemSlotNeck]: '/cata/assets/item_slots/neck.jpg',
-	[ItemSlot.ItemSlotShoulder]: '/cata/assets/item_slots/shoulders.jpg',
-	[ItemSlot.ItemSlotBack]: '/cata/assets/item_slots/shirt.jpg',
-	[ItemSlot.ItemSlotChest]: '/cata/assets/item_slots/chest.jpg',
-	[ItemSlot.ItemSlotWrist]: '/cata/assets/item_slots/wrists.jpg',
-	[ItemSlot.ItemSlotHands]: '/cata/assets/item_slots/hands.jpg',
-	[ItemSlot.ItemSlotWaist]: '/cata/assets/item_slots/waist.jpg',
-	[ItemSlot.ItemSlotLegs]: '/cata/assets/item_slots/legs.jpg',
-	[ItemSlot.ItemSlotFeet]: '/cata/assets/item_slots/feet.jpg',
-	[ItemSlot.ItemSlotFinger1]: '/cata/assets/item_slots/finger.jpg',
-	[ItemSlot.ItemSlotFinger2]: '/cata/assets/item_slots/finger.jpg',
-	[ItemSlot.ItemSlotTrinket1]: '/cata/assets/item_slots/trinket.jpg',
-	[ItemSlot.ItemSlotTrinket2]: '/cata/assets/item_slots/trinket.jpg',
-	[ItemSlot.ItemSlotMainHand]: '/cata/assets/item_slots/mainhand.jpg',
-	[ItemSlot.ItemSlotOffHand]: '/cata/assets/item_slots/offhand.jpg',
-	[ItemSlot.ItemSlotRanged]: '/cata/assets/item_slots/ranged.jpg',
-};
-export function getEmptySlotIconUrl(slot: ItemSlot): string {
-	return emptySlotIcons[slot];
-}
-
-type ItemListType = Item | Enchant | Gem | ReforgeData | ItemRandomSuffix;
-enum ItemListSortBy {
-	EP,
-	ILVL,
-}
-
-export class ItemList<T extends ItemListType> {
-	private listElem: HTMLElement;
-	private readonly simUI: SimUI;
-	private readonly player: Player<any>;
-	public id: string;
-	public label: string;
-	private slot: ItemSlot;
-	private itemData: Array<ItemData<T>>;
-	private itemsToDisplay: Array<number>;
-	private currentFilters: DatabaseFilters;
-	private searchInput: HTMLInputElement;
-	private socketColor: GemColor;
-	private computeEP: (item: T) => number;
-	private equippedToItemFn: (equippedItem: EquippedItem | null) => T | null | undefined;
-	private gearData: GearData;
-	private tabContent: Element;
-	private onItemClick: (itemData: ItemData<T>) => void;
-	private scroller: Clusterize;
-
-	private sortBy = ItemListSortBy.ILVL;
-	private sortDirection = SortDirection.DESC;
-
-	constructor(
-		id: string,
-		parent: HTMLElement,
-		simUI: SimUI,
-		currentSlot: ItemSlot,
-		currentTab: SelectorModalTabs,
-		player: Player<any>,
-		label: string,
-		gearData: GearData,
-		itemData: Array<ItemData<T>>,
-		socketColor: GemColor,
-		computeEP: (item: T) => number,
-		equippedToItemFn: (equippedItem: EquippedItem | null) => T | null | undefined,
-		onRemove: (eventID: EventID) => void,
-		onItemClick: (itemData: ItemData<T>) => void,
-	) {
-		this.id = id;
-		this.label = label;
-		this.simUI = simUI;
-		this.player = player;
-		this.itemData = itemData;
-		this.socketColor = socketColor;
-		this.computeEP = computeEP;
-		this.equippedToItemFn = equippedToItemFn;
-		this.onItemClick = onItemClick;
-
-		this.slot = currentSlot;
-		this.gearData = gearData;
-		this.currentFilters = this.player.sim.getFilters();
-
-		const selected = label === currentTab;
-		const itemLabel = label === SelectorModalTabs.Reforging ? 'Reforge' : 'Item';
-
-		const sortByIlvl = (event: MouseEvent) => {
-			event.preventDefault();
-			this.sort(ItemListSortBy.ILVL);
-		};
-		const sortByEP = (event: MouseEvent) => {
-			event.preventDefault();
-			this.sort(ItemListSortBy.EP);
-		};
-
-		const searchRef = ref<HTMLInputElement>();
-		const epButtonRef = ref<HTMLButtonElement>();
-		const filtersButtonRef = ref<HTMLButtonElement>();
-		const showEpValuesRef = ref<HTMLDivElement>();
-		const phaseSelectorRef = ref<HTMLDivElement>();
-		const matchingGemsRef = ref<HTMLDivElement>();
-		const show1hWeaponRef = ref<HTMLDivElement>();
-		const show2hWeaponRef = ref<HTMLDivElement>();
-		const modalListRef = ref<HTMLUListElement>();
-		const removeButtonRef = ref<HTMLButtonElement>();
-		const compareLabelRef = ref<HTMLElement>();
-
-		this.tabContent = (
-			<div id={this.id} className={`selector-modal-tab-pane tab-pane fade ${selected ? 'active show' : ''}`}>
-				<div className="selector-modal-filters">
-					<input ref={searchRef} className="selector-modal-search form-control" type="text" placeholder="Search..." />
-					{label === SelectorModalTabs.Items && (
-						<button ref={filtersButtonRef} className="selector-modal-filters-button btn btn-primary">
-							Filters
-						</button>
-					)}
-					<div ref={phaseSelectorRef} className="selector-modal-phase-selector" />
-					<div ref={show1hWeaponRef} className="sim-input selector-modal-boolean-option selector-modal-show-1h-weapons hide" />
-					<div ref={show2hWeaponRef} className="sim-input selector-modal-boolean-option selector-modal-show-2h-weapons hide" />
-					<div ref={matchingGemsRef} className="sim-input selector-modal-boolean-option selector-modal-show-matching-gems" />
-					<div ref={showEpValuesRef} className="sim-input selector-modal-boolean-option selector-modal-show-ep-values" />
-					<button ref={removeButtonRef} className="selector-modal-remove-button btn btn-danger">
-						Unequip Item
-					</button>
-				</div>
-				<div className="selector-modal-list-labels">
-					<span className="item-label">
-						<small>{itemLabel}</small>
-					</span>
-					{label === SelectorModalTabs.Items && (
-						<>
-							<label className="source-label">
-								<small>Source</small>
-							</label>
-							<label className="ilvl-label interactive" onclick={sortByIlvl}>
-								<small>ILvl</small>
-							</label>
-						</>
-					)}
-					<span className="ep-label interactive" onclick={sortByEP}>
-						<small>EP</small>
-						<i className="fa-solid fa-plus-minus fa-2xs"></i>
-						<button ref={epButtonRef} className="btn btn-link p-0 ms-1">
-							<i className="far fa-question-circle fa-lg"></i>
-						</button>
-					</span>
-					<span className="favorite-label"></span>
-					<span ref={compareLabelRef} className="compare-label hide"></span>
-				</div>
-				<ul ref={modalListRef} className="selector-modal-list"></ul>
-			</div>
-		);
-
-		parent.appendChild(this.tabContent);
-
-		if (this.label === SelectorModalTabs.Items) {
-			this.bindToggleCompare(compareLabelRef.value!);
-		}
-
-		tippy(epButtonRef.value!, {
-			content: EP_TOOLTIP,
-		});
-
-		if (
-			label === SelectorModalTabs.Items &&
-			(currentSlot === ItemSlot.ItemSlotMainHand || (currentSlot === ItemSlot.ItemSlotOffHand && player.getClass() === Class.ClassWarrior))
-		) {
-			if (show1hWeaponRef.value) makeShow1hWeaponsSelector(show1hWeaponRef.value, player.sim);
-			if (show2hWeaponRef.value) makeShow2hWeaponsSelector(show2hWeaponRef.value, player.sim);
-		}
-
-		if (showEpValuesRef.value) makeShowEPValuesSelector(showEpValuesRef.value, player.sim);
-
-		if (matchingGemsRef.value) {
-			makeShowMatchingGemsSelector(matchingGemsRef.value, player.sim);
-			if (!label.startsWith('Gem')) {
-				matchingGemsRef.value?.classList.add('hide');
-			}
-		}
-
-		if (phaseSelectorRef.value) makePhaseSelector(phaseSelectorRef.value, player.sim);
-
-		if (label === SelectorModalTabs.Items) {
-			const filtersMenu = new FiltersMenu(parent, player, currentSlot);
-			filtersButtonRef.value?.addEventListener('click', () => filtersMenu.open());
-		}
-
-		this.listElem = modalListRef.value!;
-		this.itemsToDisplay = [];
-
-		this.scroller = new Clusterize(
-			{
-				getNumberOfRows: () => {
-					return this.itemsToDisplay.length;
-				},
-				generateRows: (startIdx, endIdx) => {
-					const items = [];
-					for (let i = startIdx; i < endIdx; ++i) {
-						if (i >= this.itemsToDisplay.length) break;
-						items.push(this.createItemElem({ idx: this.itemsToDisplay[i], data: this.itemData[this.itemsToDisplay[i]] }));
-					}
-					return items;
-				},
-			},
-			{
-				rows: [],
-				scroll_elem: this.listElem,
-				content_elem: this.listElem,
-				item_height: 56,
-				show_no_data_row: false,
-				no_data_text: '',
-				tag: 'li',
-				rows_in_block: 16,
-				blocks_in_cluster: 2,
-			},
-		);
-
-		const removeButton = removeButtonRef.value;
-		if (removeButton) {
-			removeButton.addEventListener('click', _event => {
-				onRemove(TypedEvent.nextEventID());
-			});
-
-			switch (label) {
-				case SelectorModalTabs.Enchants:
-					removeButton.textContent = 'Remove Enchant';
-					break;
-				case SelectorModalTabs.Reforging:
-					removeButton.textContent = 'Remove Reforge';
-					break;
-				case SelectorModalTabs.Gem1:
-				case SelectorModalTabs.Gem2:
-				case SelectorModalTabs.Gem3:
-					removeButton.textContent = 'Remove Gem';
-					break;
-			}
-		}
-
-		this.updateSelected();
-
-		this.searchInput = searchRef.value!;
-		this.searchInput.addEventListener('input', () => this.applyFilters());
-	}
-
-	public sizeRefresh() {
-		this.scroller.refresh(true);
-		this.applyFilters();
-	}
-
-	public dispose() {
-		this.scroller.dispose();
-	}
-
-	private getUpdateType(item: ItemListType | null | undefined) {
-		if (!item) return null;
-		const itemProperties = Object.keys(item);
-		if ('reforge' in item && !!item.reforge) return 'reforge';
-		else if ('enchantType' in item) return 'enchant';
-		else if ('color' in item) return 'gem';
-		else if (itemProperties.length === 3 && 'name' in item && item.id && item.name && item.stats) return 'randomSuffix';
-		else return 'item';
-	}
-
-	private getItemIdByUpdateType(item: ItemListType | null | undefined) {
-		if (!item) return null;
-		const updateType = this.getUpdateType(item);
-		switch (updateType) {
-			case 'reforge':
-				return (item as ReforgeData).reforge!.id;
-			case 'enchant':
-				return (item as Enchant).effectId;
-			case 'item':
-			case 'gem':
-			case 'randomSuffix':
-				return (item as Item | Gem | ItemRandomSuffix).id;
-			default:
-				return null;
-		}
-	}
-
-	public updateSelected() {
-		const newEquippedItem = this.gearData.getEquippedItem();
-		const newItem = this.equippedToItemFn(newEquippedItem);
-		const newItemId = this.getItemIdByUpdateType(newItem);
-		const newEP = newItem ? this.computeEP(newItem) : 0;
-
-		this.scroller.elementUpdate(item => {
-			const idx = (item as HTMLElement).dataset.idx!;
-			const itemData = this.itemData[parseFloat(idx)];
-
-			if (itemData.id === newItemId) item.classList.add('active');
-			else item.classList.remove('active');
-
-			const epDeltaElem = item.querySelector<HTMLSpanElement>('.selector-modal-list-item-ep-delta');
-			if (epDeltaElem) {
-				epDeltaElem.textContent = '';
-				if (itemData.item) {
-					const listItemEP = this.computeEP(itemData.item);
-					if (newEP !== listItemEP) {
-						formatDeltaTextElem(epDeltaElem, newEP, listItemEP, 0);
-					}
-				}
-			}
-		});
-	}
-
-	public applyFilters() {
-		this.currentFilters = this.player.sim.getFilters();
-		let itemIdxs = new Array<number>(this.itemData.length);
-		for (let i = 0; i < this.itemData.length; ++i) {
-			itemIdxs[i] = i;
-		}
-
-		const currentEquippedItem = this.player.getEquippedItem(this.slot);
-		const newItem = this.equippedToItemFn(currentEquippedItem);
-		const type = this.getUpdateType(newItem);
-
-		if (type === 'item' || this.label === SelectorModalTabs.Items)
-			itemIdxs = this.player.filterItemData(itemIdxs, i => this.itemData[i].item as unknown as Item, this.slot);
-		else if (type === 'enchant' || this.label === SelectorModalTabs.Enchants)
-			itemIdxs = this.player.filterEnchantData(itemIdxs, i => this.itemData[i].item as unknown as Enchant, this.slot, currentEquippedItem);
-		else if (type === 'gem') itemIdxs = this.player.filterGemData(itemIdxs, i => this.itemData[i].item as unknown as Gem, this.slot, this.socketColor);
-
-		itemIdxs = itemIdxs.filter(i => {
-			const listItemData = this.itemData[i];
-
-			if (listItemData.phase > this.player.sim.getPhase()) {
-				return false;
-			}
-
-			if (!!this.searchInput.value.length) {
-				const formatQuery = (value: string) => value.toLowerCase().replaceAll(/[^a-zA-Z0-9\s]/g, '');
-
-				const searchQuery = formatQuery(this.searchInput.value).split(' ');
-				const name = formatQuery(listItemData.name.toString());
-
-				let include = true;
-				searchQuery.some(v => {
-					if (!name.includes(v)) include = false;
-				});
-				if (!include) {
-					return false;
-				}
-			}
-
-			return true;
-		});
-
-		if ([ItemSlot.ItemSlotTrinket1, ItemSlot.ItemSlotTrinket2].includes(this.slot)) {
-			// Trinket EP is weird so just sort by ilvl instead.
-			this.sortBy = ItemListSortBy.ILVL;
-		} else {
-			this.sortBy = ItemListSortBy.EP;
-		}
-
-		itemIdxs = this.sortIdxs(itemIdxs);
-
-		this.itemsToDisplay = itemIdxs;
-		this.scroller.update();
-
-		this.hideOrShowEPValues();
-	}
-
-	public sort(sortBy: ItemListSortBy) {
-		if (this.sortBy === sortBy) {
-			this.sortDirection = 1 - this.sortDirection;
-		} else {
-			this.sortDirection = SortDirection.DESC;
-		}
-		this.sortBy = sortBy;
-		this.itemsToDisplay = this.sortIdxs(this.itemsToDisplay);
-		this.scroller.update();
-	}
-
-	private sortIdxs(itemIdxs: Array<number>): number[] {
-		let sortFn = (itemA: T, itemB: T) => {
-			const first = this.sortDirection === SortDirection.DESC ? itemB : itemA;
-			const second = this.sortDirection === SortDirection.DESC ? itemA : itemB;
-			const diff = this.computeEP(first) - this.computeEP(second);
-			// if EP is same, sort by ilvl
-			if (Math.abs(diff) < 0.01) return (first as unknown as Item).ilvl - (second as unknown as Item).ilvl;
-			return diff;
-		};
-		switch (this.sortBy) {
-			case ItemListSortBy.ILVL:
-				sortFn = (itemA: T, itemB: T) => {
-					const first = this.sortDirection === SortDirection.DESC ? itemB : itemA;
-					const second = this.sortDirection === SortDirection.DESC ? itemA : itemB;
-					return (first as unknown as Item).ilvl - (second as unknown as Item).ilvl;
-				};
-				break;
-		}
-
-		return itemIdxs.sort((dataA, dataB) => {
-			const itemA = this.itemData[dataA];
-			const itemB = this.itemData[dataB];
-			if (this.isItemFavorited(itemA) && !this.isItemFavorited(itemB)) return -1;
-			if (this.isItemFavorited(itemB) && !this.isItemFavorited(itemA)) return 1;
-
-			return sortFn(itemA.item, itemB.item);
-		});
-	}
-
-	public hideOrShowEPValues() {
-		const labels = this.tabContent.querySelectorAll('.ep-label');
-		const container = this.tabContent.querySelectorAll('.selector-modal-list');
-		const show = this.player.sim.getShowEPValues();
-		const display = show ? '' : 'none';
-
-		for (const label of labels) {
-			(label as HTMLElement).style.display = display;
-		}
-
-		for (const c of container) {
-			if (show) c.classList.remove('hide-ep');
-			else c.classList.add('hide-ep');
-		}
-	}
-
-	private createItemElem(item: ItemDataWithIdx<T>): JSX.Element {
-		const itemData = item.data;
-		const itemEP = this.computeEP(itemData.item);
-
-		const equippedItem = this.equippedToItemFn(this.gearData.getEquippedItem());
-		const equippedItemID = equippedItem
-			? this.label === 'Enchants'
-				? (equippedItem as unknown as Enchant).effectId
-				: (equippedItem as unknown as Item).id
-			: 0;
-		const equippedItemEP = equippedItem ? this.computeEP(equippedItem) : 0;
-
-		const nameElem = ref<HTMLLabelElement>();
-		const anchorElem = ref<HTMLAnchorElement>();
-		const iconElem = ref<HTMLImageElement>();
-		const favoriteElem = ref<HTMLButtonElement>();
-		const favoriteIconElem = ref<HTMLElement>();
-		const compareContainer = ref<HTMLDivElement>();
-		const compareButton = ref<HTMLButtonElement>();
-
-		const listItemElem = (
-			<li className={`selector-modal-list-item ${equippedItemID === itemData.id ? 'active' : ''}`} dataset={{ idx: item.idx.toString() }}>
-				<div className="selector-modal-list-label-cell">
-					<a className="selector-modal-list-item-link" ref={anchorElem} dataset={{ whtticon: 'false' }}>
-						<img className="selector-modal-list-item-icon" ref={iconElem}></img>
-						<label className="selector-modal-list-item-name" ref={nameElem}>
-							{itemData.name}
-							{itemData.heroic && createHeroicLabel()}
-						</label>
-					</a>
-				</div>
-				{this.label === SelectorModalTabs.Items && (
-					<>
-						<div className="selector-modal-list-item-source-container">{this.getSourceInfo(itemData.item as unknown as Item, this.player.sim)}</div>
-						<div className="selector-modal-list-item-ilvl-container">{(itemData.item as unknown as Item).ilvl}</div>
-					</>
-				)}
-				{![ItemSlot.ItemSlotTrinket1, ItemSlot.ItemSlotTrinket2].includes(this.slot) && (
-					<div className="selector-modal-list-item-ep">
-						<span className="selector-modal-list-item-ep-value">
-							{itemEP < 9.95 ? itemEP.toFixed(1).toString() : Math.round(itemEP).toString()}
-						</span>
-						<span
-							className="selector-modal-list-item-ep-delta"
-							ref={e => itemData.item && equippedItemEP !== itemEP && formatDeltaTextElem(e, equippedItemEP, itemEP, 0)}
-						/>
-					</div>
-				)}
-				<div className="selector-modal-list-item-favorite-container">
-					<button className="selector-modal-list-item-favorite btn btn-link p-0" ref={favoriteElem}>
-						<i ref={favoriteIconElem} className="far fa-star fa-xl" />
-					</button>
-				</div>
-				<div ref={compareContainer} className="selector-modal-list-item-compare-container hide">
-					<button className="selector-modal-list-item-compare btn btn-link p-0" ref={compareButton}>
-						<i className="fas fa-arrow-right-arrow-left fa-xl" />
-					</button>
-				</div>
-			</li>
-		);
-
-		const favoriteTooltip = tippy(favoriteElem.value!);
-		const toggleFavoriteTooltipContent = (isFavorited: boolean) => favoriteTooltip.setContent(isFavorited ? 'Remove from favorites' : 'Add to favorites');
-
-		const toggleFavorite = (isFavorite: boolean) => {
-			const filters = this.player.sim.getFilters();
-			if (this.label === SelectorModalTabs.Items) {
-				const favId = itemData.id;
-				if (isFavorite) {
-					filters.favoriteItems.push(favId);
-				} else {
-					const favIdx = filters.favoriteItems.indexOf(favId);
-					if (favIdx !== -1) {
-						filters.favoriteItems.splice(favIdx, 1);
-					}
-				}
-			} else if (this.label === SelectorModalTabs.Items) {
-				const favId = getUniqueEnchantString(itemData.item as unknown as Enchant);
-				if (isFavorite) {
-					filters.favoriteEnchants.push(favId);
-				} else {
-					const favIdx = filters.favoriteEnchants.indexOf(favId);
-					if (favIdx !== -1) {
-						filters.favoriteEnchants.splice(favIdx, 1);
-					}
-				}
-			} else if (this.label.startsWith('Gem')) {
-				const favId = itemData.id;
-				if (isFavorite) {
-					filters.favoriteGems.push(favId);
-				} else {
-					const favIdx = filters.favoriteGems.indexOf(favId);
-					if (favIdx !== -1) {
-						filters.favoriteGems.splice(favIdx, 1);
-					}
-				}
-			}
-			favoriteElem.value!.classList.toggle('text-brand');
-			favoriteIconElem.value!.classList.toggle('fas');
-			favoriteIconElem.value!.classList.toggle('far');
-			listItemElem.dataset.fav = isFavorite.toString();
-			this.player.sim.setFilters(TypedEvent.nextEventID(), filters);
-		};
-
-		favoriteElem.value!.addEventListener('click', () => toggleFavorite(listItemElem.dataset.fav === 'false'));
-
-		const isFavorite = this.isItemFavorited(itemData);
-
-		if (isFavorite) {
-			favoriteElem.value!.classList.add('text-brand');
-			favoriteIconElem.value?.classList.add('fas');
-			listItemElem.dataset.fav = 'true';
-		} else {
-			favoriteIconElem.value?.classList.add('far');
-			listItemElem.dataset.fav = 'false';
-		}
-
-		toggleFavoriteTooltipContent(listItemElem.dataset.fav === 'true');
-
-		if (this.label === SelectorModalTabs.Items) {
-			const batchSimTooltip = tippy(compareButton.value!);
-
-			this.bindToggleCompare(compareContainer.value!);
-			const simUI = this.simUI instanceof IndividualSimUI ? this.simUI : null;
-			if (simUI) {
-				const checkHasItem = () => simUI.bt.hasItem(ItemSpec.create({ id: itemData.id }));
-				const toggleCompareButtonState = () => {
-					const hasItem = checkHasItem();
-					batchSimTooltip.setContent(hasItem ? 'Remove from Batch sim' : 'Add to Batch sim');
-					compareButton.value!.classList[hasItem ? 'add' : 'remove']('text-brand');
-				};
-
-				toggleCompareButtonState();
-				simUI.bt.itemsChangedEmitter.on(() => {
-					toggleCompareButtonState();
-				});
-
-				compareButton.value!.addEventListener('click', () => {
-					const hasItem = checkHasItem();
-					simUI.bt[hasItem ? 'removeItem' : 'addItem'](ItemSpec.create({ id: itemData.id }));
-
-					new Toast({
-						delay: 1000,
-						variant: 'success',
-						body: (
-							<>
-								<strong>{itemData.name}</strong> was {hasItem ? <>removed from the batch</> : <>added to the batch</>}.
-							</>
-						),
-					});
-					// TODO: should we open the bulk sim UI or should we run in the background showing progress, and then sort the items in the picker?
-				});
-			}
-		}
-
-		anchorElem.value!.addEventListener('click', (event: Event) => {
-			event.preventDefault();
-			if (event.target === favoriteElem.value) return false;
-			this.onItemClick(itemData);
-		});
-
-		itemData.actionId.fill().then(filledId => {
-			filledId.setWowheadHref(anchorElem.value!);
-			iconElem.value!.src = filledId.iconUrl;
-		});
-
-		setItemQualityCssClass(nameElem.value!, itemData.quality);
-
-		return listItemElem;
-	}
-
-	private isItemFavorited(itemData: ItemData<T>): boolean {
-		if (this.label === SelectorModalTabs.Items) {
-			return this.currentFilters.favoriteItems.includes(itemData.id);
-		} else if (this.label === 'Enchants') {
-			return this.currentFilters.favoriteEnchants.includes(getUniqueEnchantString(itemData.item as unknown as Enchant));
-		} else if (this.label.startsWith('Gem')) {
-			return this.currentFilters.favoriteGems.includes(itemData.id);
-		}
-		return false;
-	}
-
-	private getSourceInfo(item: Item, sim: Sim): JSX.Element {
-		const makeAnchor = (href: string, inner: string | JSX.Element) => {
-			return (
-				<a href={href} target="_blank" dataset={{ whtticon: 'false' }}>
-					<small>{inner}</small>
-				</a>
-			);
-		};
-
-		if (!item.sources?.length) {
-			if (item.randomSuffixOptions.length) {
-				return makeAnchor(`${ActionId.makeItemUrl(item.id)}#dropped-by`, 'World Drop');
-			} else if (isPVPItem(item)) {
-				const season = getPVPSeasonFromItem(item);
-				if (!season) return <></>;
-
-				return makeAnchor(
-					ActionId.makeItemUrl(item.id),
-					<span>
-						{season}
-						<br />
-						PVP
-					</span>,
-				);
-			}
-			return <></>;
-		}
-
-		let source = item.sources[0];
-		if (source.source.oneofKind === 'crafted') {
-			const src = source.source.crafted;
-
-			if (src.spellId) {
-				return makeAnchor(ActionId.makeSpellUrl(src.spellId), professionNames.get(src.profession) ?? 'Unknown');
-			}
-			return makeAnchor(ActionId.makeItemUrl(item.id), professionNames.get(src.profession) ?? 'Unknown');
-		} else if (source.source.oneofKind === 'drop') {
-			const src = source.source.drop;
-			const zone = sim.db.getZone(src.zoneId);
-			const npc = sim.db.getNpc(src.npcId);
-			if (!zone) {
-				throw new Error('No zone found for item: ' + item);
-			}
-
-			const category = src.category ? ` - ${src.category}` : '';
-			if (npc) {
-				return makeAnchor(
-					ActionId.makeNpcUrl(npc.id),
-					<span>
-						{zone.name} ({difficultyNames.get(src.difficulty) ?? 'Unknown'})
-						<br />
-						{npc.name + category}
-					</span>,
-				);
-			} else if (src.otherName) {
-				return makeAnchor(
-					ActionId.makeZoneUrl(zone.id),
-					<span>
-						{zone.name}
-						<br />
-						{src.otherName}
-					</span>,
-				);
-			}
-			return makeAnchor(ActionId.makeZoneUrl(zone.id), zone.name);
-		} else if (source.source.oneofKind === 'quest' && source.source.quest.name) {
-			const src = source.source.quest;
-			return makeAnchor(
-				ActionId.makeQuestUrl(src.id),
-				<span>
-					Quest
-					{item.factionRestriction === UIItem_FactionRestriction.ALLIANCE_ONLY && (
-						<img src="/cata/assets/img/alliance.png" className="ms-1" width="15" height="15" />
-					)}
-					{item.factionRestriction === UIItem_FactionRestriction.HORDE_ONLY && (
-						<img src="/cata/assets/img/horde.png" className="ms-1" width="15" height="15" />
-					)}
-					<br />
-					{src.name}
-				</span>,
-			);
-		} else if ((source = item.sources.find(source => source.source.oneofKind === 'rep') ?? source).source.oneofKind === 'rep') {
-			const factionNames = item.sources
-				.filter(source => source.source.oneofKind === 'rep')
-				.map(source =>
-					source.source.oneofKind === 'rep' ? REP_FACTION_NAMES[source.source.rep.repFactionId] : REP_FACTION_NAMES[RepFaction.RepFactionUnknown],
-				);
-			const src = source.source.rep;
-			const npcId = REP_FACTION_QUARTERMASTERS[src.repFactionId];
-			return makeAnchor(
-				ActionId.makeNpcUrl(npcId),
-				<>
-					{factionNames.map(name => (
-						<span>
-							{name}
-							{item.factionRestriction === UIItem_FactionRestriction.ALLIANCE_ONLY && (
-								<img src="/cata/assets/img/alliance.png" className="ms-1" width="15" height="15" />
-							)}
-							{item.factionRestriction === UIItem_FactionRestriction.HORDE_ONLY && (
-								<img src="/cata/assets/img/horde.png" className="ms-1" width="15" height="15" />
-							)}
-							<br />
-						</span>
-					))}
-					<span>{REP_LEVEL_NAMES[src.repLevel]}</span>
-				</>,
-			);
-		} else if (isPVPItem(item)) {
-			const season = getPVPSeasonFromItem(item);
-			if (!season) return <></>;
-
-			return makeAnchor(
-				ActionId.makeItemUrl(item.id),
-				<span>
-					{season}
-					<br />
-					PVP
-				</span>,
-			);
-		} else if (source.source.oneofKind === 'soldBy') {
-			const src = source.source.soldBy;
-			return makeAnchor(
-				ActionId.makeNpcUrl(src.npcId),
-				<span>
-					Sold by
-					<br />
-					{src.npcName}
-				</span>,
-			);
-		}
-		return <></>;
-	}
-
-	private bindToggleCompare(element: Element) {
-		const toggleCompare = () => element.classList[!this.player.sim.getShowExperimental() ? 'add' : 'remove']('hide');
-		toggleCompare();
-		this.player.sim.showExperimentalChangeEmitter.on(() => {
-			toggleCompare();
-		});
 	}
 }
